# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "金丝雀蜜罐 CanaryGuard AntiFraud",
    "description": "蜜罐反诈双形态主控：REST API 同一内核，WebUI/MCP 均为客户端（D1）。",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/speech-patterns": {
      "get": {
        "tags": [
          "speech-patterns"
        ],
        "summary": "List Patterns",
        "description": "分页列表：category 精确过滤 + keyword 模糊搜索（pattern/regex 包含匹配），含 hit_count。",
        "operationId": "list_patterns_api_v1_speech_patterns_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Patterns Api V1 Speech Patterns Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "speech-patterns"
        ],
        "summary": "Create Pattern",
        "description": "新建词条：pattern 唯一（UNIQUE(pattern,category)）、weight 1-10、category 枚举。",
        "operationId": "create_pattern_api_v1_speech_patterns_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatternCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Pattern Api V1 Speech Patterns Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/speech-patterns/{pattern_id}": {
      "put": {
        "tags": [
          "speech-patterns"
        ],
        "summary": "Update Pattern",
        "description": "编辑词条（敏感操作）：confirm=true + reason≥20，写 gate_logs(pattern.update, before/after)。",
        "operationId": "update_pattern_api_v1_speech_patterns__pattern_id__put",
        "parameters": [
          {
            "name": "pattern_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pattern Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatternUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Pattern Api V1 Speech Patterns  Pattern Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "speech-patterns"
        ],
        "summary": "Delete Pattern",
        "description": "删除词条（敏感操作）：confirm=true + reason≥20，写 gate_logs(pattern.delete)。",
        "operationId": "delete_pattern_api_v1_speech_patterns__pattern_id__delete",
        "parameters": [
          {
            "name": "pattern_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pattern Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Pattern Api V1 Speech Patterns  Pattern Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/speech-patterns/{pattern_id}/toggle": {
      "post": {
        "tags": [
          "speech-patterns"
        ],
        "summary": "Toggle Pattern",
        "description": "启停切换：停用（enabled=0）后 rule_match 实时查询不再命中；启用即时恢复（热更新）。",
        "operationId": "toggle_pattern_api_v1_speech_patterns__pattern_id__toggle_post",
        "parameters": [
          {
            "name": "pattern_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pattern Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Toggle Pattern Api V1 Speech Patterns  Pattern Id  Toggle Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/speech-patterns/import": {
      "post": {
        "tags": [
          "speech-patterns"
        ],
        "summary": "Import Patterns",
        "description": "批量导入（幂等 UPSERT）：INSERT OR IGNORE 按 UNIQUE(pattern,category) 去重，可重复执行。\n\n返回 {imported, skipped}；写 gate_logs(pattern.import) 审计（import 幂等，非敏感闸控）。",
        "operationId": "import_patterns_api_v1_speech_patterns_import_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__api__speech_patterns__ImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Import Patterns Api V1 Speech Patterns Import Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/speech-patterns/export": {
      "get": {
        "tags": [
          "speech-patterns"
        ],
        "summary": "Export Patterns",
        "description": "全量导出（备份/迁移）：返回全部词条（含 id/pattern/regex/weight/category/source/enabled/hit_count）。",
        "operationId": "export_patterns_api_v1_speech_patterns_export_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Export Patterns Api V1 Speech Patterns Export Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmRequest": {
        "properties": {
          "confirm": {
            "type": "boolean",
            "title": "Confirm",
            "default": false
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "title": "ConfirmRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportItem": {
        "properties": {
          "pattern": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Pattern"
          },
          "regex": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Regex"
          },
          "weight": {
            "type": "number",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Weight"
          },
          "category": {
            "type": "string",
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "pattern",
          "weight",
          "category"
        ],
        "title": "ImportItem"
      },
...
      "PatternCreate": {
        "properties": {
          "pattern": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Pattern"
          },
          "regex": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Regex"
          },
          "weight": {
            "type": "number",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Weight",
            "description": "权重 1-10"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled",
            "description": "缺省 1（启用）"
          }
        },
        "type": "object",
        "required": [
          "pattern",
          "weight",
          "category"
        ],
        "title": "PatternCreate"
      },
      "PatternUpdate": {
        "properties": {
          "pattern": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Pattern"
          },
          "regex": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Regex"
          },
          "weight": {
            "type": "number",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Weight",
            "description": "权重 1-10"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "confirm": {
            "type": "boolean",
            "title": "Confirm",
            "default": false
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "pattern",
          "weight",
          "category"
        ],
        "title": "PatternUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__api__speech_patterns__ImportRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ImportItem"
            },
            "type": "array",
            "maxItems": 5000,
            "minItems": 1,
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ImportRequest"
      },
...
    }
  }
}